requirements.txt
# ThreatGPT Requirements
# Generated from pyproject.toml dependencies

# Core Framework Dependencies
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
pyyaml>=6.0.1,<7.0.0
click>=8.1.7,<9.0.0
rich>=13.7.0,<14.0.0

# AI/LLM Provider Dependencies
openai>=1.3.0,<2.0.0
anthropic>=0.7.0,<0.8.0

# Database Dependencies
asyncpg>=0.29.0,<0.30.0
redis>=5.0.1,<6.0.0
sqlalchemy[asyncio]>=2.0.23,<3.0.0
alembic>=1.13.0,<2.0.0

# HTTP and Networking
httpx>=0.25.2,<0.26.0

# Template and Rendering
jinja2>=3.1.2,<4.0.0

# Logging and Monitoring
structlog>=23.2.0,<24.0.0
prometheus-client>=0.19.0,<0.20.0

# File Handling
aiofiles>=23.2.1,<24.0.0

# Authentication and Security
python-multipart>=0.0.6,<0.0.7
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0

# Reliability and Retry Logic
tenacity>=8.2.3,<9.0.0

# Development Dependencies (Optional - install with: pip install -r requirements-dev.txt)
# pytest>=7.4.0,<8.0.0
# pytest-asyncio>=0.21.0,<0.22.0
# pytest-cov>=4.1.0,<5.0.0
# pytest-mock>=3.12.0,<4.0.0
# black>=23.11.0,<24.0.0
# flake8>=6.1.0,<7.0.0
# mypy>=1.7.0,<2.0.0
# pre-commit>=3.6.0,<4.0.0
# isort>=5.13.0,<6.0.0
# bandit>=1.7.5,<2.0.0
# safety>=2.3.5,<3.0.0
# coverage>=7.3.2,<8.0.0